requirements.txt
requests==2.24.0
pillow==9.2.0
transformers==4.21.0
numpy==1.20.0
flask==2.0.1